move picker and order monitor prioritization into single file
